update services & models into latest version.

+++ src/services/expert-base.js
@@ -1,8 +1,9 @@
 /**
  * Created by zhanglimin on 17/9/1.
  */
-import { request, config } from '../utils';
 import { sysconfig } from 'systems';
+import { request, config, nextAPI } from 'utils';
+import { apiBuilder, F, H, Action } from 'utils/next-api-builder';
 
 const { api } = config;
 
@@ -18,24 +19,28 @@ export async function getExpert(offset, size) {
 }
 
 export async function getExpertDetail(id, offset, size) {
-  return request(
-    api.getExpertDetailList
+  return request(api.getExpertDetailList
       .replace(':ebid', id)
       .replace(':offset', offset)
       .replace(':size', size),
-    { method: 'GET' },
-  );
+
+    {
+      method: 'GET',
+    });
 }
 
 export async function rosterManage({ payload }) {
   const { id, name, email, perm } = payload;
-  return request(
-    api.invokeRoster.replace(':id', id),
+  return request(api.invokeRoster
+      .replace(':id', id),
     {
       method: 'PUT',
-      body: JSON.stringify({ name, email, perm }),
-    },
-  );
+      body: JSON.stringify({
+        name,
+        email,
+        perm,
+      }),
+    });
 }
 
 export async function addExpertBase({ title, desc, pub }) {
@@ -53,14 +58,26 @@ export async function addExpertBase({ title, desc, pub }) {
 
 export async function addExpertToEB({ payload }) {
   const { ebid, aids } = payload;
-  return request(api.addExpertToEB.replace(':ebid', ebid),
-    {
-      method: 'PUT',
-      body: JSON.stringify({
-        aids,
-      }),
-    });
+  if (sysconfig.USE_NEXT_EXPERT_BASE_SEARCH) {
+    const opts = [];
+    for (let i = 0; i < ebid.length; i++) {
+      opts.push({ aid: aids[0], eid: ebid[i], opt: 'add' });
+    }
+    const nextapi = apiBuilder.create(Action.person_eb.alter, 'person_eb.alter')
+      .param({ opts });
+
+    return nextAPI({ data: [nextapi.api] });
+  } else {
+    return request(api.addExpertToEB.replace(':ebid', ebid[0]),
+      {
+        method: 'PUT',
+        body: JSON.stringify({
+          aids,
+        }),
+      });
+  }
 }
+
 
 export async function deleteByKey(key) {
   return request(
@@ -70,10 +87,21 @@ export async function deleteByKey(key) {
 }
 
 export async function removeExpertsFromEBByPid({ pid, rid }) {
-  return request(
-    api.removeExpertsFromEBByPid.replace(':rid', rid).replace(':pid', pid),
-    { method: 'DELETE' },
-  );
+  if (sysconfig.USE_NEXT_EXPERT_BASE_SEARCH) {
+    const opts = [];
+    for (let i = 0; i < rid.length; i++) {
+      opts.push({ aid: pid, eid: rid[i], opt: 'remove' });
+    }
+    const nextapi = apiBuilder.create(Action.person_eb.alter, 'person_eb.alter')
+      .param({ opts });
+
+    return nextAPI({ data: [nextapi.api] });
+  } else {
+    return request(
+      api.removeExpertsFromEBByPid.replace(':rid', rid[0]).replace(':pid', pid),
+      { method: 'DELETE' },
+    );
+  }
 }
 
 // export async function searchExpert({ payload }) {
@@ -103,4 +131,82 @@ export async function updateToBProfileExtra(aid, extra) {
       method: 'PATCH',
       body: JSON.stringify({ extra }),
     });
+}
+
+export async function getExpertList() {
+  const data = [
+    {
+      desc: 'adfs',
+      id: '5a62e2cae33a5d27239e9f00',
+      logo: 'aa/bb/fgh',
+      name_zh: 'zhangsan111',
+      names: [
+        'bson',
+      ],
+      score: 0,
+      sourcetype: '',
+      type: 'generated',
+    },
+    {
+      desc: 'adfs',
+      id: '5a62e2cae33a5d27239e9f01',
+      logo: 'aa/bb/fgh',
+      name_zh: 'zhangsan222',
+      names: [
+        'bson',
+      ],
+      score: 0,
+      sourcetype: '',
+      type: 'generated',
+    },
+    {
+      desc: 'adfs1111111',
+      id: '5a62e2cae33a5d27239e9f02',
+      logo: 'aa/bb/fgh',
+      name_zh: 'zhangsan33333',
+      names: [
+        'bson',
+      ],
+      score: 0,
+      sourcetype: '',
+      type: 'generated',
+    },
+    {
+      desc: 'adfs2222222',
+      id: '5a62e2cae33a5d27239e9f03',
+      logo: 'aa/bb/fgh',
+      name_zh: 'zhangsan4444',
+      names: [
+        'bson',
+      ],
+      score: 0,
+      sourcetype: '',
+      type: 'generated',
+    },
+    {
+      desc: 'adfs',
+      id: '5a62e36ae33a5d27239e9f31',
+      logo: 'aa/bb/fgh',
+      name_zh: 'zhangsan333',
+      parents: [
+        '5a62e2cae33a5d27239e9f00',
+      ],
+      score: 0,
+      sourcetype: '',
+      type: 'generated',
+    },
+    {
+      desc: 'adfs111',
+      id: '5a62e36ae33a5d27239e9f32',
+      logo: 'aa/bb/fgh',
+      name_zh: 'zhangsan444',
+      parents: [
+        '5a62e2cae33a5d27239e9f00',
+      ],
+      score: 0,
+      sourcetype: '',
+      type: 'generated',
+    },
+  ];
+  return data;
 }
